+++ src/apis/loginAPI.js
@@ -41,17 +41,16 @@ export const dupCheck = async (data) => {
   return answer;
 };
 
-//로그아웃 
+//로그아웃
 export const logout = async (refreshToken) => {
   const answer = { result: true };
   try {
-    const res = await axios.default.delete("/auth/signout", {
+    await axios.default.delete("/auth/signout", {
       data: {
-        refreshToken
-      }
+        refreshToken,
+      },
     });
-    console.log(res)
-    localStorage.clear()
+    localStorage.clear();
   } catch (err) {
     answer.result = false;
     answer.msg = err.response;
