+++ src/components/page/login/index.js
@@ -1,10 +1,11 @@
 import Indroduce from "./Introduce";
 import LoginPage from "./LoginPage";
 import SignUpPage2 from "./SignUpPage2";
+import SignUpPage3 from "./SignUpPage3";
 import Choice from "./Choice";
 import FindPassword from "./FindPassword";
 import FindId from "./FindId";
 import Line from "./Line";
 import SignUpPage from "./SignUpPage";
 
-export { Indroduce, LoginPage, SignUpPage, Choice, FindId, FindPassword, Line, SignUpPage2 };
+export { Indroduce, LoginPage, SignUpPage, SignUpPage3, Choice, FindId, FindPassword, Line, SignUpPage2 };
